+++ app/src/main/java/com/chaquo/python/TestThread.java
@@ -2,6 +2,8 @@ package com.chaquo.python;
 
 public class TestThread {
 
+    public static boolean blocked = false;
+
     public static class BlockingConstructor {
         public BlockingConstructor(long delay) {
             BlockingMethods.blockStatic(delay);
@@ -16,9 +18,12 @@ public class TestThread {
 
         public static void blockStatic(long delay) {
             try {
+                blocked = true;
                 Thread.sleep(delay);
             } catch (InterruptedException e) {
                 throw new RuntimeException(e);
+            } finally {
+                blocked = false;
             }
         }
     }