+++ app/src/main/java/com/chaquo/python/TestThreadAppClass.java
@@ -1,0 +1,7 @@
+package com.chaquo.python;
+
+// To avoid interference from the class cache, this class should not be referenced from anywhere
+// except test_app_class in test_thread.
+public class TestThreadAppClass {
+    public static String STR = "hello world";
+}